unable to navigate to new addCoin page
solution: add new addcoin page to addCoin.js and addCoin.html #1280
