+++ imports/ui/pages/addCoin/smart.wizard.custom.css
@@ -1,3 +1,4 @@
+@@ -1,221 +0,0 @@
 /*!
  * SmartWizard v4.3.x
  * jQuery Wizard Plugin
